refactor(inputters): Move XML parser setup local to reader
